Java Study Function
