+++ Java/javaBasic/src/main/java/com/example/javabasic/FunctionAndLambda/App.java
@@ -1,7 +1,6 @@
-package com.example.javabasic.ch01;
+package com.example.javabasic.FunctionAndLambda;
 
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.function.UnaryOperator;
