+++ Java/javaBasic/src/main/java/com/example/javabasic/FunctionAndLambda/RunSomething.java
@@ -1,0 +1,19 @@
+package com.example.javabasic.FunctionAndLambda;
+
+
+@FunctionalInterface
+public interface RunSomething {
+
+    void doIt();
+
+//
+//    static void printName(){
+//        System.out.println("minsoo");
+//    }
+//
+//    default void printAge() {
+//        System.out.println("26");
+//    }
+}
+
+// 함수형 인터페이스 하나의 추상 메서드만 존재
