--- Java/javaBasic/src/main/java/com/example/javabasic/ch00/customer/Grade.java
@@ -1,5 +1,0 @@
-package com.example.javabasic.ch00.customer;
-
-public enum Grade {
-    VIP , BASIC
-}
