--- Java/javaBasic/src/main/java/com/example/javabasic/ch00/menu/drink/Temperature.java
@@ -1,5 +1,0 @@
-package com.example.javabasic.ch00.menu.drink;
-
-public enum Temperature {
-    HOT,ICE
-}
