--- Java/javaBasic/src/main/java/com/example/javabasic/ch01/RunSomething2.java
@@ -1,19 +1,0 @@
-package com.example.javabasic.ch01;
-
-
-@FunctionalInterface
-public interface RunSomething2 {
-
-    int doIt(int number);
-
-
-    static void printName(){
-        System.out.println("minsoo");
-    }
-
-    default void printAge() {
-        System.out.println("26");
-    }
-}
-
-// 함수형 인터페이스 하나의 추상 메서드만 존재
