+++ Java/javaBasic/src/main/java/com/example/javabasic/object/customer/Grade.java
@@ -1,0 +1,5 @@
+package com.example.javabasic.object.customer;
+
+public enum Grade {
+    VIP , BASIC
+}
