+++ Java/javaBasic/src/main/java/com/example/javabasic/object/menu/drink/Temperature.java
@@ -1,0 +1,5 @@
+package com.example.javabasic.object.menu.drink;
+
+public enum Temperature {
+    HOT,ICE
+}
